addded

+++ python_introduction/finance_calculator.py
@@ -1,7 +1,7 @@
-Monthly_income = int(input("Enter your monthly income:"))
-Monthly_expense = int(input("Enter your total monthly expenses:"))
-Monthly_Savings = Monthly_income - Monthly_expense
-Projected_Savings = int(Monthly_Savings * 12 + (Monthly_Savings * 12 * 0.05))
+Monthly_income = float(input("Enter your monthly income:"))
+Monthly_expense = float(input("Enter your total monthly expenses:"))
+Monthly_Savings =float( Monthly_income) - float(Monthly_expense)
+Projected_Savings = float(Monthly_Savings * 12 + (Monthly_Savings * 12 * 0.05))
 print("Your monthly savings are $",Monthly_Savings)
 print("Projected savings after one year, with interest, is: $",Projected_Savings)
